requirements/base.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Data Processing
pandas>=2.0.0,<2.1.0
numpy>=1.24.0,<1.27.0
scipy>=1.11.0
openpyxl>=3.1.0
pyarrow>=14.0.0

# Data Profiling
ydata-profiling>=4.6.0

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0
optuna>=3.4.0
imbalanced-learn>=0.11.0

# Explainability
shap>=0.43.0

# Computer Vision
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
transformers>=4.30.0
pillow>=10.0.0
opencv-python>=4.8.0
grad-cam>=1.4.0

# Experiment Tracking
mlflow>=2.8.0

# Frontend
streamlit>=1.28.0

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
